Merge fixup: update from operator-framework/rukpak
This copies everything from operator-framework/rukpak, which really ends
up being just the `Makefile` and `Dockerfile`. Both the `Dockerfile` and
the `Makefile` have been tweaked to work with upstream changes.

Versions of the locally built components previously in `/bin`, now in
`hack/tools/bin`, are based on the `go.mod` in the new location rather
than versions in the `Makefile`.

Signed-off-by: Todd Short <todd.short@me.com>